refactor: side -> team 폴더명 수정
