+++ src/main/java/janggi/domain/piece/LeapingPiece.java
@@ -3,7 +3,7 @@ package janggi.domain.piece;
 import janggi.domain.Delta;
 import janggi.domain.MovePath;
 import janggi.domain.Position;
-import janggi.domain.side.TeamType;
+import janggi.domain.team.TeamType;
 
 import java.util.List;
