+++ src/test/java/janggi/domain/piece/LeapingPieceTest.java
@@ -1,6 +1,6 @@
 package janggi.domain.piece;
 
-import janggi.domain.side.TeamType;
+import janggi.domain.team.TeamType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
